requirements.txt
pyserial
ultralytics
requests
paho-mqtt
